Debugging VI

+++ socket_server/server_socket.c
@@ -70,19 +70,19 @@ static void signal_handler(int signal)
 void func(int connfd)
 {
     int bytes_sent, package_count = 1;
-    char buff[sizeof(float) + sizeof(float) + 20] = {0};
+    char buff[sizeof(double) + sizeof(double) + 20] = {0};
     char toClient[50] = {0};
     unsigned int priority;
-    float temperature_data,humidity_data;
+    double temperature_data,humidity_data;
     // infinite loop for chat
     //while(1) 
     //{
-	if(mq_receive(mqd, buff, (sizeof(float) + sizeof(float)), &priority) == -1)
+	if(mq_receive(mqd, buff, (sizeof(double) + sizeof(double)), &priority) == -1)
 	{
 	    printf("\n\rError in receiving message from the queue. Error: %s", strerror(errno));
 	}
-	memcpy(&temperature_data, buff, sizeof(float));
-	memcpy(&humidity_data, buff + sizeof(float), sizeof(float));
+	memcpy(&temperature_data, buff, sizeof(double));
+	memcpy(&humidity_data, buff + sizeof(double), sizeof(double));
 	
 	sprintf(toClient, "Temperature = %0.2lf", temperature_data);
 	sprintf(toClient, "Humidity = %0.2lf", humidity_data);
